refactor: move replaceObjectBodyRegexTool to service layer
